first/serializers.py
from rest_framework import serializers
from first.models import Recipe

class RecipeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Recipe
        fields = ('name','estimated_time')