requirements.txt
argon2-cffi==19.2.0
bcrypt==3.1.7
cffi==1.13.2
dj-database-url==0.5.0
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.11.0
Faker==2.0.4
gunicorn==20.0.4
Markdown==3.1.1
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
whitenoise==5.0.1
